Version 3227: * add search by published Date (to search over position) * add handling of import categories (pinned still to test) * do not delete old logs on import (we have the correct LogId now) * give focus lost event for EditTextFields (so that we can react on changed input, still not perfect)

+++ CB_UI/src/CB_UI/GL_UI/Main/Actions/CB_Action_LoadLogs.java
@@ -79,7 +79,8 @@ public class CB_Action_LoadLogs extends CB_Action {
 
                         Iterator<LogEntry> iterator = logList.iterator();
                         LogDAO dao = new LogDAO();
-                        dao.deleteLogs(GlobalCore.getSelectedCache().Id);
+                        if (loadAllLogs)
+                            dao.deleteLogs(GlobalCore.getSelectedCache().Id);
                         do {
                             ChangedCount++;
                             try {
